requirements.txt
beautifulsoup4==4.12.3
inflect==7.0.0
matplotlib==3.8.0
networkx==3.2.1
nltk==3.8.1
numpy<2.0.0
pandas==2.2.2
Pillow==10.3.0
Requests==2.32.3
scikit_learn==1.4.1.post1
scipy==1.14.0
sentence_transformers==2.6.1
SoundsLike==0.0.11
SoundsLike==0.0.11
spacy==3.7.4
statsmodels==0.14.2
tiktoken==0.7.0
torch==2.2.2
tqdm==4.66.2
transformers==4.39.1
umap==0.1.1
wordcloud==1.9.3
wordfreq==3.1.1
openai~=0.28.0
seaborn~=0.13.2
scikit-learn~=1.4.1.post1
replicate~=0.31.0
pysat~=3.2.0